refactor(checker): move type-literal duplicate-property scan into modifier module
The arch guard caps checker src files at 2000 LOC; after adding the TS2687
modifier-agreement call, `core.rs` reached 2004 lines. Relocate
`check_type_literal_duplicate_properties` into the new
`duplicate_property_modifiers` module so the duplicate-property scan and the
modifier-agreement helper live together. Behavior unchanged; core.rs drops to
1883 lines.
